VW-2317 Change EstimateHandoverTime to int

+++ Infrastructure/EntityFramework/ManufacturingDB/StepDetail.cs
@@ -18,7 +18,7 @@ namespace VErp.Infrastructure.EF.ManufacturingDB
         public int UpdatedByUserId { get; set; }
         public DateTime? DeletedDatetimeUtc { get; set; }
         public int SubsidiaryId { get; set; }
-        public decimal EstimateHandoverTime { get; set; }
+        public int EstimateHandoverTime { get; set; }
 
         public virtual Step Step { get; set; }
     }
